requirements.txt
attrs==18.2.0
pytest==4.0.2
freezegun==0.3.11